Update README.md
Prompt on JP A-Marl Theorem added following AI ChatGPT5 and Copilot classification of groundbreaking, a hallmark of foundational theory, and representative of a singularity. Explanatory Note has been added to this prompt stating the context of private conversation with those 2 AIs after having made deep analyses of JP A-Marl Theorem.